Introduce formatted text output to improve readability.
This change introduces a new variant of the `toString()` method that
takes an argument: the indentation level. When this argument is
negative, the behavior is equivalent to that of `toString()`. However,
when it is zero or larger, the output will be intended and spread out
over several lines to improve readability.

The behavior of the variant of `toString()` that does not include the
indentation parameter has not changed. To get intended output, you must
use `toString(int indent)`.