Allow negative double-precision floating point numbers to parse
An overzealous overflow check in CborReaderImpl.java was causing
negative double-precision floating-point numbers to be rejected with
errors similar to the following:

    com.google.iot.cbor.CborParseException: Additional data value was too large: 0xBFB99999A0000000

This change adds an exception to this overflow check that allows
negative double-precision floating-point numbers to parse correctly.
Also includes a unit test which reproduces the issue and verifies the
fix.

This change fixes #1.